requirements.txt
pandas==1.0.5
lxml==4.5.1
mysql.connector==2.2.9
Jinja2==2.11.2